fix(datastore): nil pointer exception
